fix(simple-select): remove with: auto for transparent
