+++ src/components/select/multi-select/multi-select.spec.js
@@ -2,7 +2,10 @@ import React, { useRef } from "react";
 import { act } from "react-dom/test-utils";
 import { mount } from "enzyme";
 
-import { testStyledSystemMargin } from "../../../__spec_helper__/test-utils";
+import {
+  assertStyleMatch,
+  testStyledSystemMargin,
+} from "../../../__spec_helper__/test-utils";
 import MultiSelect from "./multi-select.component";
 import Textbox from "../../../__experimental__/components/textbox";
 import SelectTextbox from "../select-textbox/select-textbox.component";
@@ -11,6 +14,7 @@ import SelectList from "../select-list/select-list.component";
 import { StyledSelectList } from "../select-list/select-list.style";
 import Pill from "../../pill";
 import Label from "../../../__experimental__/components/label";
+import InputPresentationStyle from "../../../__experimental__/components/input/input-presentation.style";
 
 describe("MultiSelect", () => {
   testStyledSystemMargin((props) => getSelect(props));
@@ -96,6 +100,21 @@ describe("MultiSelect", () => {
       );
       expect(wrapper.find(SelectList).exists()).toBe(true);
     });
+  });
+
+  it.each([
+    ["small", "32px"],
+    ["medium", "40px"],
+    ["large", "48px"],
+  ])("the input toggle icon should have proper left margin", (a, expected) => {
+    const wrapper = renderSelect({ size: a });
+    assertStyleMatch(
+      {
+        paddingRight: expected,
+      },
+      wrapper,
+      { modifier: `${InputPresentationStyle}` }
+    );
   });
 
   describe('when the "value" prop is passed', () => {
