+++ src/components/select/multi-select/multi-select.style.js
@@ -1,6 +1,10 @@
 import styled from "styled-components";
 import StyledPill from "../../pill/pill.style";
 import { baseTheme } from "../../../style/themes";
+import InputIconToggleStyle from "../../../__experimental__/components/input-icon-toggle/input-icon-toggle.style";
+import StyledSelect from "../select.style";
+import InputPresentationStyle from "../../../__experimental__/components/input/input-presentation.style";
+import sizes from "../../../__experimental__/components/input/input-sizes.style";
 
 const StyledSelectPillContainer = styled.div`
   display: flex;
@@ -13,8 +17,21 @@ const StyledSelectPillContainer = styled.div`
   }
 `;
 
+const StyledSelectMultiSelect = styled(StyledSelect)`
+  ${InputIconToggleStyle} {
+    margin-right: 0;
+    position: absolute;
+    right: 0;
+    height: 100%;
+  }
+
+  ${InputPresentationStyle} {
+    padding-right: ${({ size }) => sizes[size].height};
+  }
+`;
+
 StyledSelectPillContainer.defaultProps = {
   theme: baseTheme,
 };
 
-export default StyledSelectPillContainer;
+export { StyledSelectPillContainer, StyledSelectMultiSelect };
